bc práce

See merge request teaching/nprg045/tuma/performance-unit-test-evaluator!24

+++ implementation/PerfEval/src/main/java/cz/cuni/mff/d3s/perfeval/init/InitCommand.java
@@ -16,52 +16,61 @@ import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
-import java.time.Duration;
+
+import static cz.cuni.mff.d3s.perfeval.command.SetupUtilities.GIT_FILE_NAME;
 
 /**
- * Command for initializing perfeval in directory
+ * Command for initializing perfeval in directory.
  */
 public class InitCommand implements Command {
     /**
-     * String representation of key for critical value in ini file
+     * String representation of key for critical value in ini file.
      */
-    private static final String CRIT_VALUE_KEY = "critValue.key";
+    private static final String CRIT_VALUE_KEY = "statistic.falseAlarmProbability";
     /**
-     * String representation of key for maximal confidence interval width in ini file
+     * String representation of key for maximal confidence interval width in ini file.
      */
-    private static final String MAX_CI_WIDTH_KEY = "maxCIWidth.key";
+    private static final String MAX_CI_WIDTH_KEY = "statistic.accuracy";
     /**
-     * String representation of key for maximal time on test in ini file
+     * String representation of key for maximal time on test in ini file.
      */
-    private static final String MAX_TIME_KEY = "maxTime.key";
+    private static final String MAX_TEST_COUNT_KEY = "statistic.maxTestCount";
     /**
-     * String representation of key for version in ini file
+     * String representation of key for minimal test count in ini file.
      */
-    private static final String VERSION_KEY = "version.key";
+    private static final String MIN_TEST_COUNT_KEY = "statistic.minTestCount";
     /**
-     * String representation of key for git presence in ini file
+     * String representation of key for tolerance in ini file.
      */
-    private static final String GIT_PRESENCE_KEY = "git.key";
+    private static final String TOLERANCE_KEY = "statistic.tolerance";
     /**
-     * String representation of key for parser name in ini file
+     * String representation of key for high run demands in ini file.
      */
-    private static final String PARSER_NAME_KEY = "parserName.key";
+    private static final String HIGH_RUN_DEMAND_ALERT_KEY = "alerts.highDemandOfRuns";
     /**
-     * String representation of key for tolerance in ini file
+     * String representation of key for improved performance in ini file.
      */
-    private static final String TOLERANCE_KEY = "tolerance.key";
+    private static final String IMPROVED_PERFORMANCE_ALERT_KEY = "alerts.improvedPerformance";
     /**
-     * String representation of true in ini file
+     * String representation of key for git presence in ini file.
+     */
+    private static final String GIT_PRESENCE_KEY = "project.git";
+    /**
+     * String representation of key for parser name in ini file.
+     */
+    private static final String PARSER_NAME_KEY = "project.parserName";
+    /**
+     * String representation of true in ini file.
      */
     private static final String TRUE_STRING = "TRUE";
     /**
-     * String representation of false in ini file
+     * String representation of false in ini file.
      */
     private static final String FALSE_STRING = "FALSE";
 
     /**
      * Returns default config if perfeval is not initialized in this directory
-     * Otherwise returns config from ini file
+     * otherwise returns config from ini file.
      *
      * @param iniFilePath path to ini file
      * @return config from ini file or default config
@@ -81,39 +90,61 @@ public class InitCommand implements Command {
     }
 
     /**
-     * Returns true if perfeval is initialized in this directory
+     * Returns true if perfeval is initialized in this directory.
      *
-     * @param perfevalDirPath path to perfeval directory
+     * @param iniFilePath path to ini file
      * @return true if perfeval is initialized in this directory, false otherwise
      */
-    public static boolean isPerfevalInitializedInThisDirectory(Path perfevalDirPath) {
-        return perfevalDirPath.toFile().isDirectory();
+    private static boolean isPerfevalInitializedInThisDirectory(Path iniFilePath) {
+        return iniFilePath.toFile().exists();
     }
 
     /**
-     * Reads config from ini file
+     * Reads config from ini file.
      *
      * @param iniFilePath path to ini file
      * @return config from ini file
      * @throws ConfigurationException         if ini file cannot be read
      * @throws PerfEvalInvalidConfigException if config (instance of PerfEvalConfig) cannot be created
      */
-    public static PerfEvalConfig readFromIniFile(Path iniFilePath) throws ConfigurationException, PerfEvalInvalidConfigException {
+    private static PerfEvalConfig readFromIniFile(Path iniFilePath) throws ConfigurationException, PerfEvalInvalidConfigException {
         Configurations configs = new Configurations();
         INIConfiguration config = configs.ini(iniFilePath.toFile());
         double critValue = Double.parseDouble(config.getString(CRIT_VALUE_KEY));
         double maxCIWidth = Double.parseDouble(config.getString(MAX_CI_WIDTH_KEY));
-        Duration timeUnit = Duration.ofNanos(Long.parseLong(config.getString(MAX_TIME_KEY)));
-        String version = config.getString(VERSION_KEY);
+        int maxTestCount = config.getInt(MAX_TEST_COUNT_KEY);
+        int minTestCount = config.getInt(MIN_TEST_COUNT_KEY);
         boolean gitPresence = config.getString(GIT_PRESENCE_KEY).compareTo(TRUE_STRING) == 0;
         String parserName = config.getString(PARSER_NAME_KEY);
         MeasurementParser parser = ParserFactory.getParser(parserName);
         double tolerance = Double.parseDouble(config.getString(TOLERANCE_KEY));
-        return new PerfEvalConfig(gitPresence, timeUnit, maxCIWidth, critValue, version, parser, tolerance);
+        boolean highRunDemandsAlert = config.getString(HIGH_RUN_DEMAND_ALERT_KEY).compareTo(TRUE_STRING) == 0;
+        boolean improvedPerformanceAlert = config.getString(IMPROVED_PERFORMANCE_ALERT_KEY).compareTo(TRUE_STRING) == 0;
+
+        if(critValue <= 0 || critValue >= 1){
+            throw new PerfEvalInvalidConfigException("Critical value must be in (0, 1)");
+        }
+        if(maxCIWidth <= 0 || maxCIWidth >= 1){
+            throw new PerfEvalInvalidConfigException("Maximal width of confidence interval must be in (0, 1)");
+        }
+        if(minTestCount < 0) {
+            throw new PerfEvalInvalidConfigException("Minimal count of tests must be non-negative.");
+        }
+        if(minTestCount > maxTestCount) {
+            throw new PerfEvalInvalidConfigException("Minimal count of tests must be less than maximal count of tests.");
+        }
+        if(maxTestCount <= 0){
+            throw new PerfEvalInvalidConfigException("Maximal count of tests must be positive");
+        }
+        if(tolerance < 0 || tolerance > 1){
+            throw new PerfEvalInvalidConfigException("Tolerance must be in [0, 1]");
+        }
+
+        return new PerfEvalConfig(gitPresence, minTestCount, maxTestCount, maxCIWidth, critValue, parser, tolerance, highRunDemandsAlert, improvedPerformanceAlert);
     }
 
     /**
-     * Creates ini file
+     * Creates ini file.
      *
      * @param iniFilePath    path to ini file
      * @param perfevalConfig config to be saved to ini file
@@ -121,21 +152,27 @@ public class InitCommand implements Command {
      * @throws IOException            if ini file cannot be created
      */
     private static void createIniFile(Path iniFilePath, PerfEvalConfig perfevalConfig) throws ConfigurationException, IOException {
-        if (iniFilePath.toFile().exists() && !iniFilePath.toFile().delete())
+        if (iniFilePath.toFile().exists() && !iniFilePath.toFile().delete()) {
             throw new IOException("ini file cannot be deleted");
+        }
+
+        Path gitFilePath = iniFilePath.getParent().getParent().resolve(GIT_FILE_NAME);
+        boolean isThereAGitFile = gitFilePath.toFile().exists();
 
         INIConfiguration config = new INIConfiguration();
 
         // Set INI properties
         config.setProperty(CRIT_VALUE_KEY, perfevalConfig.getCritValue());
         config.setProperty(MAX_CI_WIDTH_KEY, perfevalConfig.getMaxCIWidth());
-        config.setProperty(MAX_TIME_KEY, perfevalConfig.getMaxTimeOnTest().getNano());
-        String gitPresenceString = perfevalConfig.hasGitFilePresence() ? TRUE_STRING : FALSE_STRING;
+        config.setProperty(MIN_TEST_COUNT_KEY, perfevalConfig.getMinTestCount());
+        config.setProperty(MAX_TEST_COUNT_KEY, perfevalConfig.getMaxTestCount());
+        String gitPresenceString = perfevalConfig.hasGitFilePresence() || isThereAGitFile ? TRUE_STRING : FALSE_STRING;
         config.setProperty(GIT_PRESENCE_KEY, gitPresenceString);
-        config.setProperty(VERSION_KEY, perfevalConfig.getVersion());
-        //TODO: solve parser name
-        //config.setProperty(PARSER_NAME_KEY, perfevalConfig.getMeasurementParser().getParserName());
+        //solve parser name --> method toString()
+        config.setProperty(PARSER_NAME_KEY, perfevalConfig.getMeasurementParser().toString());
         config.setProperty(TOLERANCE_KEY, perfevalConfig.getTolerance());
+        config.setProperty(HIGH_RUN_DEMAND_ALERT_KEY, perfevalConfig.isHighRunDemandAlert() ? TRUE_STRING : FALSE_STRING);
+        config.setProperty(IMPROVED_PERFORMANCE_ALERT_KEY, perfevalConfig.isImprovedPerformanceAlert() ? TRUE_STRING : FALSE_STRING);
 
         // Save to the INI file
         FileWriter writer = new FileWriter(iniFilePath.toFile());
@@ -144,65 +181,53 @@ public class InitCommand implements Command {
     }
 
     /**
-     * true if perfeval should be initialized even if it is already initialized in this directory, false otherwise
+     * True if perfeval should be initialized even if it is already initialized in this directory, false otherwise.
      */
     final boolean forceFlag;
     /**
-     * path to perfeval directory
+     * Path to perfeval directory.
      */
     final Path perfevalDirPath;
     /**
-     * path to gitignore file
+     * Path to gitignore file.
      */
     final Path gitIgnorePath;
     /**
-     * path to ini file
+     * Path to ini file.
      */
     final Path iniFilePath;
     /**
-     * paths to files that should be created
-     */
-    final Path[] emptyFilesToCreate;
-    /**
-     * paths to files that should be ignored by git
-     */
-    final Path[] gitIgnoredFiles;
-    /**
-     * config to be saved to ini file
+     * Config to be saved to ini file.
      */
     final PerfEvalConfig config;
 
     /**
-     * Constructor for InitCommand
+     * Constructor for InitCommand.
      *
      * @param perfevalDirPath    path to perfeval directory
      * @param gitIgnorePath      path to gitignore file
      * @param iniFilePath        path to ini file
-     * @param emptyFilesToCreate paths to files that should be created
-     * @param gitIgnoredFiles    paths to files that should be ignored by git
      * @param config             config to be saved to ini file
      * @param forceFlag          true if perfeval should be initialized even if it is already initialized in this directory, false otherwise
      */
-    public InitCommand(Path perfevalDirPath, Path gitIgnorePath, Path iniFilePath, Path[] emptyFilesToCreate,
-                       Path[] gitIgnoredFiles, PerfEvalConfig config, boolean forceFlag) {
+    public InitCommand(Path perfevalDirPath, Path gitIgnorePath, Path iniFilePath,
+                       PerfEvalConfig config, boolean forceFlag) {
         this.gitIgnorePath = gitIgnorePath;
         this.iniFilePath = iniFilePath;
-        this.emptyFilesToCreate = emptyFilesToCreate;
         this.config = config;
         this.perfevalDirPath = perfevalDirPath;
-        this.gitIgnoredFiles = gitIgnoredFiles;
         this.forceFlag = forceFlag;
     }
 
     /**
-     * Executes init command
+     * Executes init command.
      *
      * @return exit code, OK if init was successful, notInitialized if perfeval is already initialized in this directory and force flag is not set
      * @throws PerfEvalCommandFailedException if init fails
      */
     @Override
     public ExitCode execute() throws PerfEvalCommandFailedException {
-        if (isPerfevalInitializedInThisDirectory(perfevalDirPath)) {
+        if(forceFlag && isPerfevalInitializedInThisDirectory(perfevalDirPath)) {
             try {
                 deleteDirectory(perfevalDirPath);
             } catch (IOException e) {
@@ -210,8 +235,8 @@ public class InitCommand implements Command {
                 exception.initCause(e);
                 throw exception;
             }
-        } else if (!forceFlag) {
-            throw new PerfEvalCommandFailedException(ExitCode.notInitialized);
+        } else if (isPerfevalInitializedInThisDirectory(perfevalDirPath)) {
+            return ExitCode.alreadyInitialized;
         }
         try {
             createPerfEvalFiles();
@@ -224,7 +249,7 @@ public class InitCommand implements Command {
     }
 
     /**
-     * Creates perfeval files
+     * Creates perfeval files.
      *
      * @throws IOException if perfeval files cannot be created
      */
@@ -236,42 +261,26 @@ public class InitCommand implements Command {
         } catch (ConfigurationException e) {
             throw new IOException("Config file cannot be created", e);
         }
-        createGitIgnoreFile(this.gitIgnorePath, gitIgnoredFiles);
-        for (Path emptyFile : emptyFilesToCreate) {
-            createEmptyFile(emptyFile);
-        }
+        createGitIgnoreFile(this.gitIgnorePath);
     }
 
     /**
-     * Creates empty file
-     *
-     * @param emptyFilePath path to empty file to be created
-     * @throws IOException if empty file cannot be created
-     */
-    private static void createEmptyFile(Path emptyFilePath) throws IOException {
-        if (!emptyFilePath.toFile().createNewFile()) {
-            throw new IOException("File " + emptyFilePath.getFileName() + " was not created");
-        }
-    }
-
-    /**
-     * Creates gitignore file
+     * Creates gitignore file.
      *
      * @param gitIgnoreFilePath path to gitignore file to be created
-     * @param ignoredFiles      paths to files that should be ignored by git
      * @throws IOException if gitignore file cannot be created
      */
-    private static void createGitIgnoreFile(Path gitIgnoreFilePath, Path[] ignoredFiles) throws IOException {
+    private static void createGitIgnoreFile(Path gitIgnoreFilePath) throws IOException {
         BufferedWriter writer = Files.newBufferedWriter(gitIgnoreFilePath, StandardCharsets.UTF_8, StandardOpenOption.CREATE);
-        for (Path ignoredFile : ignoredFiles) {
-            writer.write(ignoredFile.getFileName().toString());
-            writer.newLine();
-        }
+        writer.write("# Ignore perfeval files");
+        writer.newLine();
+        writer.write("*");
+        writer.newLine();
         writer.close();
     }
 
     /**
-     * Deletes directory
+     * Deletes directory.
      *
      * @param dirPath path to directory to be deleted
      * @throws IOException if directory cannot be deleted
@@ -279,8 +288,9 @@ public class InitCommand implements Command {
     private static void deleteDirectory(Path dirPath) throws IOException {
         if (Files.isDirectory(dirPath)) {
             DirectoryStream<Path> directoryStream = Files.newDirectoryStream(dirPath);
-            for (Path entry : directoryStream)
+            for (Path entry : directoryStream) {
                 deleteDirectory(entry);
+            }
             directoryStream.close();
         }
         Files.delete(dirPath);
